src/main/java/com/blazartech/droolsdemo/simple/TaxiFareCalculatorService.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.blazartech.droolsdemo.simple;

import org.kie.api.runtime.KieContainer;
import org.kie.api.runtime.KieSession;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Service;

/**
 *
 * @author AAR1069
 */
@Service
public class TaxiFareCalculatorService {
    
    private static final Logger logger = LoggerFactory.getLogger(TaxiFareCalculatorService.class);
    
    @Autowired
    @Qualifier("taxiFareRulesContainer")
    private KieContainer kieContainer;
 
    public Long calculateFare(TaxiRide taxiRide, Fare rideFare) {
        KieSession kieSession = kieContainer.newKieSession();
        kieSession.setGlobal("rideFare", rideFare);
        kieSession.setGlobal("logger", logger);
        kieSession.insert(taxiRide);
        kieSession.fireAllRules();
        kieSession.dispose();
        return rideFare.getRideFare();
    }
}
